Move LogPayload to separate file
